refactor(ios): split Matrix runtime concerns
